@@ -1,8 +1,11 @@
 module github.com/hotmall/commandline
 
-go 1.13
+go 1.19
+
+require github.com/stretchr/testify v1.8.4
 
 require (
-	github.com/smartystreets/assertions v1.0.1 // indirect
-	github.com/smartystreets/goconvey v1.6.4
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
